refactor: change structure of the project (DD-15)
