--- src/app/features/TrainA/pages/home-page/home-page.component.html
@@ -1,1 +1,0 @@
-<p>home-page works!</p>
