--- src/app/features/TrainA/pages/home-page/home-page.component.ts
@@ -1,11 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-  selector: 'dd-home-page',
-  standalone: true,
-  imports: [],
-  templateUrl: './home-page.component.html',
-  styleUrl: './home-page.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class HomePageComponent {}
