--- src/app/features/home/components/search-form/search-form.component.scss
@@ -1,8 +1,0 @@
-@import '@taiga-ui/core/styles/taiga-ui-local.scss';  
-.submit {  
-  flex: 0 1 10rem;  
-
-  @media #{$tui-mobile} {  
-  flex: 1 1 10rem;  
-}  
-}
