+++ src/app/features/search/interfaces/city-coordinates.interface.ts
