+++ src/app/features/search/interfaces/city-info.interface.ts
