+++ src/app/features/search/interfaces/search-route-params.interface.ts
