+++ src/app/features/search/interfaces/search-route-response.interface.ts
