+++ src/app/features/search/pages/home-page/home-page.component.scss
