+++ src/app/features/search/services/search-api/search-api.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
-import { SearchRouteParams } from '@features/home/interfaces/search-route-params.interface';
+import { SearchRouteParams } from '@features/search/interfaces/search-route-params.interface';
 
 @Injectable({
   providedIn: 'root',
